requirements.txt
fastapi
uvicorn
llama-cpp-python
huggingface-hub
jinja2
pymupdf
pytesseract
pillow
python-multipart